Add files via upload
This project provides a comprehensive solution for remotely waking up devices on a network using the Wake-on-LAN (WoL) protocol. It includes functionality to check device status, create and encrypt WoL magic packets, and send these packets securely over a network.
Basically if you are in Vpn network this script may work.